requirements.txt
Django>=5.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
python-dotenv>=1.0.0
openai>=1.10.0
anthropic>=0.8.0
google-generativeai>=0.3.0
